Move starship detection into StarshipManager.kt
This means Starship.kt is just a class for holding data now
