+++ src/main/kotlin/starships/StarshipManager.kt
@@ -1,7 +1,11 @@
 package io.github.petercrawley.minecraftstarshipplugin.starships
 
+import io.github.petercrawley.minecraftstarshipplugin.MinecraftStarshipPlugin
 import io.github.petercrawley.minecraftstarshipplugin.MinecraftStarshipPlugin.Companion.getPlugin
+import io.github.petercrawley.minecraftstarshipplugin.MinecraftStarshipPlugin.Companion.mainConfig
+import io.github.petercrawley.minecraftstarshipplugin.customblocks.MSPMaterial
 import org.bukkit.Bukkit
+import org.bukkit.ChunkSnapshot
 import org.bukkit.Material
 import org.bukkit.block.Block
 import org.bukkit.block.data.BlockData
@@ -73,11 +77,73 @@ object StarshipManager: BukkitRunnable() {
 
 	fun getStarshipAt(block: Block, requester: Player): Starship { return Starship(block, requester) }
 
-	fun activateStarship(starship: Starship, requester: Player) {
-		if (starship.owner == requester) {
-			starship.pilot = requester
+	fun activateStarship(starship: Starship) {
+		activeStarships.add(starship)
+	}
 
-			activeStarships.add(starship)
-		}
+	// TODO: If we ever need more speed... https://en.wikipedia.org/wiki/Flood_fill#Span_Filling
+	fun detectStarship(starship: Starship) {
+		starship.pilot?.sendMessage("Detecting Starship.")
+
+		Bukkit.getScheduler().runTaskAsynchronously(getPlugin(), Runnable {
+			val chunkStorage = mutableMapOf<MSPChunkLocation, ChunkSnapshot>()
+
+			val checkedBlocks = mutableSetOf<MSPBlockLocation>() // List of blocks we have checked
+			val blocksToCheck = mutableSetOf<MSPBlockLocation>() // List of blocks we need to check
+
+			blocksToCheck.addAll(starship.detectedBlocks) // We need to check that all the blocks we already know about
+
+			// Construct the undetectable list
+			val undetectables = mutableSetOf<MSPMaterial>()
+			undetectables.addAll(MinecraftStarshipPlugin.forcedUndetectable)  // Add all forced undetectables
+			undetectables.addAll(MinecraftStarshipPlugin.defaultUndetectable) // Add all default undetectables
+			undetectables.removeAll(starship.allowedBlocks)                   // Remove all that have been allowed by the user
+
+			// Get the detection limit from the config file.
+			val detectionLimit = mainConfig.getInt("detectionLimit", 500000)
+
+			while (blocksToCheck.isNotEmpty()) {
+				if (starship.detectedBlocks.size > detectionLimit) {
+					starship.pilot?.sendMessage("Reached arbitrary detection limit. ($detectionLimit)")
+					return@Runnable
+				}
+
+				// Get and remove the first item
+				val currentBlock = blocksToCheck.first()
+				blocksToCheck.remove(currentBlock)
+
+				val chunkCoord = MSPChunkLocation(currentBlock)
+
+				val chunk = chunkStorage.getOrPut(chunkCoord) {
+					currentBlock.world.getChunkAt(chunkCoord.x, chunkCoord.z).chunkSnapshot // FIXME: Potentially not thread-safe
+				}
+
+				val type = MSPMaterial(chunk.getBlockData(currentBlock.x - (chunkCoord.x shl 4), currentBlock.y, currentBlock.z - (chunkCoord.z shl 4)))
+
+				if (undetectables.contains(type)) continue
+
+				starship.detectedBlocks.add(currentBlock)
+
+				// FIXME: This method of making a set and then iterating over it is probably slower then just 6 if statements.
+				// List of neighbouring blocks.
+				mutableSetOf(
+						currentBlock.relative( 1, 0, 0),
+						currentBlock.relative(-1, 0, 0),
+						currentBlock.relative( 0, 1, 0),
+						currentBlock.relative( 0,-1, 0),
+						currentBlock.relative( 0, 0, 1),
+						currentBlock.relative( 0, 0,-1)
+
+						// If it's not a block we have checked, check it
+				).forEach {
+					if (!checkedBlocks.contains(it)) {
+						checkedBlocks.add(it)
+						blocksToCheck.add(it)
+					}
+				}
+			}
+
+			starship.pilot?.sendMessage("Detected " + starship.detectedBlocks.size + " blocks.") // FIXME: Potentially not thread-safe
+		})
 	}
 }
